refact reducers and store
